parseMemos export and support for all transactions
