+++ src/v1/ledger/parse/account-delete.ts
@@ -1,7 +1,7 @@
 import * as assert from 'assert'
 import {removeUndefined} from '../../common'
 import {classicAddressToXAddress} from 'ripple-address-codec'
-import {parseMemos} from './utils'
+import parseMemos from "./memos";
 
 export type FormattedAccountDelete = {
   // account (address) of an account to receive any leftover XRP after deleting the sending account.
