+++ src/v1/ledger/parse/amendment.ts
@@ -1,7 +1,11 @@
+import { removeUndefined } from "../../common";
+import parseMemos from "./memos";
+
 function parseAmendment(tx: any) {
-  return {
-    amendment: tx.Amendment
-  }
+  return removeUndefined({
+    amendment: tx.Amendment,
+    memos: parseMemos(tx),
+  });
 }
 
-export default parseAmendment
+export default parseAmendment;
