+++ src/v1/ledger/parse/check-cancel.ts
@@ -1,6 +1,6 @@
 import * as assert from 'assert'
 import {removeUndefined} from '../../common'
-import {parseMemos} from './utils'
+import parseMemos from "./memos";
 
 export type FormattedCheckCancel = {
   // ID of the Check ledger object to cancel.
