+++ src/v1/ledger/parse/escrow-cancellation.ts
@@ -1,6 +1,6 @@
 import * as assert from 'assert'
-import {parseMemos} from './utils'
 import {removeUndefined} from '../../common'
+import parseMemos from "./memos";
 
 function parseEscrowCancellation(tx: any): object {
   assert.ok(tx.TransactionType === 'EscrowCancel')
