+++ src/v1/ledger/parse/payment-channel-fund.ts
@@ -1,7 +1,8 @@
 import * as assert from 'assert'
-import {parseTimestamp,parseMemos} from './utils'
+import {parseTimestamp} from './utils'
 import {removeUndefined} from '../../common'
 import parseRippledAmount from './ripple-amount'
+import parseMemos from "./memos";
 
 function parsePaymentChannelFund(tx: any): object {
   assert.ok(tx.TransactionType === 'PaymentChannelFund')
